requirements-dev.txt
-r requirements.txt

flake8==7.1.1
black==24.10.0
isort==5.13.2

pytest==8.3.3
pytest-django==4.8.0
